fix(cli): use minimal enrollment helpers instead of full module
The full enrollments_sections_logic.go file triggered unused-code lint
failures because its command entrypoints are not yet committed. Keep only
the enrollment types and fetch helper needed by groups collaboration.
